# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OctoManger API v2
  version: 2.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v2/agents:
    get:
      operationId: listAgents
      responses:
        "200":
          description: Agent inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreateInput"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /api/v2/agents/{id}:
    get:
      operationId: getAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: patchAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPatchInput"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted agent
  /api/v2/agents/{id}/status:
    get:
      operationId: getAgentStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent runtime status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
  /api/v2/agents/{id}/start:
    post:
      operationId: startAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Agent accepted for start
  /api/v2/agents/{id}/stop:
    post:
      operationId: stopAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Agent accepted for stop
  /api/v2/agents/{id}/events:
    get:
      operationId: streamAgentEvents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
  /api/v2/accounts:
    get:
      operationId: listAccounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreateInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /api/v2/accounts/{id}:
    get:
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    patch:
      operationId: patchAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountPatchInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted account
  /api/v2/accounts/{id}/execute:
    post:
      operationId: executeAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountExecuteInput"
      responses:
        "200":
          description: Account action execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExecuteResult"
...
components:
  schemas:
...
    Agent:
      type: object
      required:
        - id
        - name
        - plugin_key
        - action
        - input
        - desired_state
        - runtime_state
      properties:
        id:
          type: integer
        name:
          type: string
        plugin_key:
          type: string
        action:
          type: string
        input:
          type: object
          additionalProperties: true
        desired_state:
          type: string
        runtime_state:
          type: string
        last_error:
          type: string
    AgentCreateInput:
      type: object
      required:
        - name
        - plugin_key
        - action
        - input
      properties:
        name:
          type: string
        plugin_key:
          type: string
        action:
          type: string
        input:
          type: object
          additionalProperties: true
    AgentPatchInput:
      type: object
      properties:
        name:
          type: string
        input:
          type: object
          additionalProperties: true
    AgentStatus:
      type: object
      required:
        - id
        - runtime_state
        - desired_state
        - updated_at
      properties:
        id:
          type: integer
        runtime_state:
          type: string
        desired_state:
          type: string
        last_error:
          type: string
        last_heartbeat_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Account:
      type: object
      required:
        - id
        - identifier
        - status
        - tags
        - spec
      properties:
        id:
          type: integer
        account_type_id:
          type: integer
        account_type_key:
          type: string
        identifier:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        spec:
          type: object
          additionalProperties: true
    AccountCreateInput:
      type: object
      required:
        - account_type_id
        - identifier
        - status
        - tags
        - spec
      properties:
        account_type_id:
          type: integer
        identifier:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        spec:
          type: object
          additionalProperties: true
    AccountPatchInput:
      type: object
      properties:
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        spec:
          type: object
          additionalProperties: true
    AccountExecuteInput:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        params:
          type: object
          additionalProperties: true
    AccountExecuteResult:
      type: object
      required:
        - status
      properties:
        status:
          type: string
        result:
          type: object
          additionalProperties: true
        error_code:
          type: string
        error_message:
          type: string
...